Update & leverage commons.py functionalities
