+++ app/tests/conftest.py
@@ -1,10 +1,12 @@
 import pytest
 import platform
 import asyncio as io
+import pathlib as plib
 
 from .support.models import create_user
 from ..utils.db_connect import db_connector, drop_schema
-from ..settings import DB_CONFIGS
+from ..utils.commons import get_attribute_values
+from ..settings import DB_CONFIGS, APP_NAME
 from .. import models
 
 @pytest.fixture(scope='session')
@@ -24,14 +26,35 @@ def configure_db():
     conn = db_connector('test')
     yield conn
     conn.close()
+    
+def make_dir(base_path: plib.Path, sub_dir: str):
+    new_path = base_path / sub_dir
+    new_path.mkdir()
+    return new_path
+    
+@pytest.fixture(scope='module')
+def prog_path(tmp_path_factory, mocker):
+    prog = tmp_path_factory.mktemp(f'AppData/{APP_NAME}')
+    mocker.path('app.utils.working_dirs.WorkingDirs.app_data', return_value=prog)
+    return prog
+
+@pytest.fixture(scope='module')
+def assets_path(prog_path):
+    return make_dir(prog_path, 'assets')
+
+@pytest.fixture(scope='module')
+def images_path(assets_path):
+    return make_dir(assets_path, 'images')
+
+@pytest.fixture(scope='module')
+def avatars_path(images_path):
+    return make_dir(images_path, 'avatars')
 
 @pytest.fixture(scope='class', autouse=True)
 def purge_db():
     yield
-    for model_cls in map(models.__dict__.get, models.__all__):
-        if model_cls.cls_name() in ('_model', 'activity'):
-            continue
-        model_cls.delete().execute()
+    for cls in get_attribute_values(models, exceptions=('_Model', 'Activity', 'ActivityLog')):
+        cls.delete().execute()
 
 @pytest.fixture(scope='class')
 def user(request):
