+++ app/tests/controllers/conftest.py
@@ -1,18 +1,11 @@
 import pytest
 import requests
-import pathlib as plib
 
 from ..support.mocks.servers import AppServerSpec
 from ..support.mocks.clients import AppClientSpec
 from ..support.unittests import PyTestCase
 from ...utils.interfaces import AttributeDict
 from ...models import User
-from ...settings import APP_NAME
-
-def make_dir(base_path: plib.Path, sub_dir: str):
-    new_path = base_path / sub_dir
-    new_path.mkdir()
-    return new_path
 
 @pytest.fixture(scope='class')
 def server(event_loop):
@@ -31,22 +24,6 @@ def user_avatar(request):
             'content_type': resp.headers['Content-Type'],
             'content_length': int(resp.headers['Content-Length'])
         })
-
-@pytest.fixture(scope='class')
-def prog_path(tmp_path_factory):
-    return tmp_path_factory.mktemp(f'AppData/{APP_NAME}')
-
-@pytest.fixture(scope='class')
-def assets_path(prog_path):
-    return make_dir(prog_path, 'assets')
-
-@pytest.fixture(scope='class')
-def images_path(assets_path):
-    return make_dir(assets_path, 'images')
-
-@pytest.fixture(scope='class')
-def avatars_path(images_path):
-    return make_dir(images_path, 'avatars')
 
 @pytest.mark.usefixtures('server')
 class ControllerTestCases(PyTestCase):
